refactored how we are exporting the components
