+++ client/src/components/BattleAnnouncer/index.js
@@ -1,2 +1,14 @@
-// where we do exports
-export * from './BattleAnnouncer';
+// battle announcer will use typed messages to type out the actions that happen during the battle
+import { useTypedMessage } from 'hooks';
+import styles from './styles.module.css';
+
+const BattleAnnouncer = ({ message }) => {
+  const typedMessage = useTypedMessage(message);
+
+  return (
+    <div className={styles.main}>
+      <div className={styles.message}>{typedMessage}</div>
+    </div>
+  );
+};
+export default BattleAnnouncer;
