--- client/src/components/BattleMenu/BattleMenu.js
@@ -1,16 +1,0 @@
-// this is where the Attack, Magic attack, and Heal buttons will live
-import styles from './styles.module.css';
-
-export const BattleMenu = ({ onAttack, onMagic, onHeal }) => (
-  <div className={styles.main}>
-    <div onClick={onAttack} className={styles.option}>
-      Attack
-    </div>
-    <div onClick={onMagic} className={styles.option}>
-      Magic
-    </div>
-    <div onClick={onHeal} className={styles.option}>
-      Heal
-    </div>
-  </div>
-);
