+++ client/src/components/BattleMenu/Index.js
@@ -1,2 +1,17 @@
 // where we do exports
-export * from './BattleMenu';
+import styles from './styles.module.css';
+
+ const BattleMenu = ({ onAttack, onMagic, onHeal }) => (
+  <div className={styles.main}>
+    <div onClick={onAttack} className={styles.option}>
+      Attack
+    </div>
+    <div onClick={onMagic} className={styles.option}>
+      Magic
+    </div>
+    <div onClick={onHeal} className={styles.option}>
+      Heal
+    </div>
+  </div>
+);
+export default BattleMenu;
