+++ client/src/components/index.js
@@ -1,4 +1,4 @@
-export * from './App';
+// export * from './App';
 export * from './Bar';
 export * from './Battle';
 export * from './EndMenu';
